Move github project ops to github/project
