+++ src/remote/query.rs
@@ -7,7 +7,8 @@ use crate::{
     api_traits::ApiOperation,
     error,
     github::merge_request::GithubMergeRequestFields,
-    github::{GithubMemberFields, GithubPipelineFields, GithubProjectFields},
+    github::project::{GithubMemberFields, GithubProjectFields},
+    github::GithubPipelineFields,
     gitlab::cicd::GitlabPipelineFields,
     gitlab::merge_request::GitlabMergeRequestFields,
     gitlab::project::{GitlabMemberFields, GitlabProjectFields},
